Adds debug helper
This adds the debug package and debug.Helper which allows for easier
management of debug log levels. This also gets rid of the Debug field in
Config as it was really confusing on what logging that enabled, and also
firecracker-runtime.json. The log_level/LogLevel field in Config and the
firecracker-runtime.json has been renamed to log_levels and LogLevels.

Signed-off-by: xibz <impactbchang@gmail.com>